backend-node/src/modules/users/services/UpdateUserAvatarService.ts
import 'reflect-metadata';

import IUserRepository from '../repositories/IUserRepository';
import User from '../infra/typeorm/entities/User';
import { injectable, inject } from 'tsyringe';

import IStorageProvider from '@shared/container/providers/StorageProvider/models/IStorageProvider';

import AppErrors from '../../../shared/errors/AppError';

interface Request {
  user_id: string;
  avatarFilename: string;
}

@injectable()
class UpdateUserAvatarService {

  constructor(
    @inject('UserRepository')
    private usersRepository: IUserRepository,

    @inject('StorageProvider')
    private storageProvider: IStorageProvider
  ) { }


  public async execute({ user_id, avatarFilename }: Request): Promise<User> {

    const user = await this.usersRepository.findById(user_id);

    if (!user) {
      throw new AppErrors('Para mudar a imagem, favor se logar.', 401);
    }

    if (user.avatar) {

      await this.storageProvider.deleteFile(user.avatar);
    }

    const filename = await this.storageProvider.saveFile(avatarFilename);

    user.avatar = filename;

    await this.usersRepository.save(user);

    return user;

  }
}

export default UpdateUserAvatarService;